WS-1002 renamed apiClientClasses.ts to Client.ts
